refactor(app): move column slice to column folder
