+++ app/src/app/store.ts
@@ -6,8 +6,8 @@ import {
 } from "@reduxjs/toolkit";
 import { TypedUseSelectorHook, useDispatch, useSelector } from "react-redux";
 
+import column from "features/Columns/columnSlice";
 import filter from "features/Filters/filterSlice";
-import column from "features/Mappings/columnSlice";
 import mapping from "features/Mappings/mappingSlice";
 import resource from "features/Mappings/resourceSlice";
 import source from "features/Sources/sourceSlice";
