+++ app/src/services/api/endpoints.ts
@@ -112,6 +112,7 @@ export const api = generatedApi
   });
 
 export const {
+  // Sources
   useApiSourcesListQuery,
   useApiSourcesCreateMutation,
   useApiSourcesRetrieveQuery,
